Added lombok plugin and moved & merged entities
