--- src/main/java/com/app/Main.java
@@ -1,5 +1,0 @@
-package com.app;
-
-public class Main {
-
-}
