--- src/main/java/com/app/entites/Poll.java
@@ -1,29 +1,0 @@
-package com.app.entites;
-
-import lombok.Data;
-
-import javax.persistence.*;
-import java.util.Date;
-import java.util.List;
-
-@Data
-@Entity
-public class Poll {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
-
-    private String question;
-    private String alternativeA;
-    private String alternativeB;
-    private int duration;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "owner_id")
-    private User owner;
-
-    @OneToMany(fetch = FetchType.LAZY, mappedBy = "voter")
-    private List<Vote> votes;
-
-}
