+++ src/main/java/no/hvl/dat250/h2020/group5/Main.java
@@ -1,9 +1,6 @@
 package no.hvl.dat250.h2020.group5;
 
 
-import java.io.*;
-
-
 public class Main {
 
     public static void main(String[] args) {
