--- src/main/java/no/hvl/dat250/h2020/group5/entities/Voter.java
@@ -1,44 +1,0 @@
-package no.hvl.dat250.h2020.group5.entities;
-
-import javax.persistence.*;
-import java.util.ArrayList;
-import java.util.List;
-
-
-@Entity
-@Inheritance(strategy = InheritanceType.JOINED)
-public abstract class Voter {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.SEQUENCE)
-    private Integer id;
-
-    @Column(length = 20)
-    private String userName;
-
-    private List<Vote> votes = new ArrayList<>();
-
-    public Integer getId() {
-        return id;
-    }
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-    public String getUserName() {
-        return userName;
-    }
-
-    public void setUserName(String userName) {
-        this.userName = userName;
-    }
-
-    public List<Vote> getVotes() {
-        return votes;
-    }
-
-    public void setVotes(List<Vote> votes) {
-        this.votes = votes;
-    }
-}
