+++ src/main/lombok/no/hvl/dat250/h2020/group5/entities/DisplayDevice.java
@@ -1,7 +1,10 @@
 package no.hvl.dat250.h2020.group5.entities;
 
+import lombok.Data;
+
 import javax.persistence.Entity;
 
+@Data
 @Entity(name = "DisplayDevice")
 public class DisplayDevice extends Device {
 }
