+++ src/main/lombok/no/hvl/dat250/h2020/group5/entities/VotingDevice.java
@@ -1,8 +1,11 @@
 package no.hvl.dat250.h2020.group5.entities;
 
 
+import lombok.Data;
+
 import javax.persistence.Entity;
 
+@Data
 @Entity(name = "VotingDevice")
 public class VotingDevice extends Device {
 }
